refactor(home-component): refact home-component, create different sub-component
