+++ opis-manager/src/app/academic-year/academic-year.component.html
@@ -1,0 +1,53 @@
+<div class="text-center">
+  <div class="list-option-container">
+    Seleziona anno:
+    <br>
+    <select
+      class="btn btn-opis text-center switch-years"
+      [(ngModel)]="selectedYear"
+      (ngModelChange)="getSchedeOfCdsForSelectedYearAndShowAcademicYearChart()"
+    >
+      <option class="text-center">--</option>
+      <option class="text-center" *ngFor="let year of configService.config.years" [ngValue]="year"> {{ year }} </option>
+    </select>
+  </div>
+</div>
+
+<div [hidden]="!selectedYear">
+  <div>
+    <ul class="nav nav-tabs" id="tabs">
+      <li class="nav-item tabs_button"><button class="btn btn-opis" (click)="switchVal(1)" [ngClass]="switcherValues === 1 ? 'selected-tab' : ''">V1</button></li>
+      <li class="nav-item tabs_button"><button class="btn btn-opis" (click)="switchVal(2)" [ngClass]="switcherValues === 2 ? 'selected-tab' : ''">V2</button></li>
+      <li class="nav-item tabs_button"><button class="btn btn-opis" (click)="switchVal(3)" [ngClass]="switcherValues === 3 ? 'selected-tab' : ''">V3</button></li>
+      <li>
+        <div class="input-group addons-bar">
+          <div class="input-group-prepend">
+            <span class="input-group-text selected-tab" id="basic-addon1">
+              <fa-icon [icon]="faSearch" size="lg" class="opis-icon"></fa-icon>
+            </span>
+          </div>
+
+          <input
+            type="text"
+            class="form-control"
+            placeholder="Fondamenti..."
+            aria-describedby="basic-addon1"
+            [(ngModel)]="subject"
+            (change)="getSchedeOfCdsForSelectedYearAndShowAcademicYearChart()"
+          >
+        </div>
+      </li>
+    </ul>
+
+    <br>
+    <div>Colore per numerosità di schede <strong>MIN - <div class="color-bar"></div> - MAX</strong></div>
+
+    <br>
+    <app-value-details [switcher]="switcherValues"></app-value-details>
+
+    <!-- Anno Accademico canvas -->
+    <div id="v1" [hidden]="switcherValues != 1"><canvas id="v1-canvas"></canvas></div>
+    <div id="v2" [hidden]="switcherValues != 2"><canvas id="v2-canvas"></canvas></div>
+    <div id="v3" [hidden]="switcherValues != 3"><canvas id="v3-canvas"></canvas></div>
+  </div>
+</div>
