+++ opis-manager/src/app/academic-year/academic-year.component.scss
@@ -1,0 +1,12 @@
+
+.color-bar {
+  display: inline-block;
+  height: 20px;
+  width: 200px;
+  background: rgba(0, 0, 0, 0) linear-gradient(to right, rgb(255, 200, 45), rgb(0, 121, 107)) repeat scroll 0% 0%;
+  vertical-align: middle;
+}
+
+.switch-years {
+  width: 350px;
+}
