+++ opis-manager/src/app/home/home.component.scss
@@ -1,7 +1,6 @@
 $color_dark: #00695c;
 $color_light: #fff;
 $color_main: #00897b;
-$font_family_1: "Helvetica Neue";
 
 .container-switcher {
   float: left;
@@ -65,20 +64,4 @@ $font_family_1: "Helvetica Neue";
       position: absolute;
     }
   }
-}
-
-.switch-years {
-    width: 350px;
-}
-
-.color-bar {
-  display: inline-block;
-  height: 20px;
-  width: 200px;
-  background: rgba(0, 0, 0, 0) linear-gradient(to right, rgb(255, 200, 45), rgb(0, 121, 107)) repeat scroll 0% 0%;
-  vertical-align: middle;
-}
-
-#math-stats {
-  white-space: pre-wrap;
 }
