+++ opis-manager/src/app/teaching/teaching.component.spec.ts
@@ -1,25 +1,46 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { Component, EventEmitter, Input, Output } from '@angular/core';
+import { async, TestBed } from '@angular/core/testing';
+import { FormsModule } from '@angular/forms';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+import { ChartModule } from 'angular2-chartjs';
+import { ValueDetailsComponent } from '../value-details/value-details.component';
 import { TeachingComponent } from './teaching.component';
 
-describe('TeachingComponent', () => {
-  let component: TeachingComponent;
-  let fixture: ComponentFixture<TeachingComponent>;
+/* Mock NgSliderComponent */
+// tslint:disable-next-line: component-selector
+@Component({selector: 'ng5-slider', template: ''})
+export class NgSliderStubComponent {
+    @Input() value;
+    @Input() highValue;
+    @Input() options;
+    @Input() manualRefresh;
+    @Output() valueChange =  new EventEmitter<any>();
+}
 
+describe('TeachingComponent', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ TeachingComponent ]
+      declarations: [
+        TeachingComponent,
+        NgSliderStubComponent,
+        ValueDetailsComponent,
+      ],
+      imports: [
+        FontAwesomeModule,
+        HttpClientTestingModule,
+        FormsModule,
+        ChartModule,
+        NgbModule,
+      ]
     })
     .compileComponents();
   }));
 
-  beforeEach(() => {
-    fixture = TestBed.createComponent(TeachingComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
   it('should create', () => {
-    expect(component).toBeTruthy();
+    const fixture = TestBed.createComponent(TeachingComponent);
+    const app = fixture.debugElement.componentInstance;
+    expect(app).toBeTruthy();
   });
 });
